Plugins revamp (#2108)
* extentions file

* plugs folder creation
