+++ app/plugins/extensions.js
@@ -1,0 +1,18 @@
+module.exports = {
+  availableExtensions: new Set([
+    'onApp', 'onWindow', 'onRendererWindow', 'onUnload', 'middleware',
+    'reduceUI', 'reduceSessions', 'reduceTermGroups',
+    'decorateMenu', 'decorateTerm', 'decorateHyper',
+    'decorateHyperTerm', // for backwards compatibility with hyperterm
+    'decorateHeader', 'decorateTerms', 'decorateTab',
+    'decorateNotification', 'decorateNotifications',
+    'decorateTabs', 'decorateConfig', 'decorateEnv',
+    'decorateTermGroup', 'decorateSplitPane', 'getTermProps',
+    'getTabProps', 'getTabsProps', 'getTermGroupProps',
+    'mapHyperTermState', 'mapTermsState',
+    'mapHeaderState', 'mapNotificationsState',
+    'mapHyperTermDispatch', 'mapTermsDispatch',
+    'mapHeaderDispatch', 'mapNotificationsDispatch',
+    'extendKeymaps'
+  ])
+};
